Изменил вложенность, пакет db.md может быть применен к нескольким БД, выразил это в структуре
