--- components/databases/md/schemes/md/tests/Integration/entrypoints/test_alembic.py
@@ -1,7 +1,0 @@
-from db.md.md.entrypoints import alembic
-
-
-def test_make_config():
-    """Дымовой тест."""
-    config = alembic.make_config()
-    assert config
