--- components/databases/md/schemes/md/tests/Integration/test_settings.py
@@ -1,6 +1,0 @@
-from db.md.md.settings import DataBaseSettings
-
-
-def test_settings():
-    db_settings = DataBaseSettings()
-    assert isinstance(db_settings.LOGGING_CONFIG, dict)
