+++ components/db_schemes/md/src/db/md/fixtures/init/initial.sql
