+++ components/db_schemes/md/tests/Integration/entrypoints/test_alembic.py
@@ -1,0 +1,7 @@
+from db.md.entrypoints import alembic
+
+
+def test_make_config():
+    """Дымовой тест."""
+    config = alembic.make_config()
+    assert config
