+++ components/db_schemes/md/tests/Integration/fixtures/test_tools.py
@@ -1,0 +1,5 @@
+from db.md.fixtures import tools
+
+
+def test_create_conn():
+    assert tools.create_conn()
