+++ components/db_schemes/md/tests/Integration/test_settings.py
@@ -1,0 +1,6 @@
+from db.md.settings import DataBaseSettings
+
+
+def test_settings():
+    db_settings = DataBaseSettings()
+    assert isinstance(db_settings.LOGGING_CONFIG, dict)
