+++ components/db_schemes/mes/pyproject.toml
@@ -1,0 +1,7 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "db.mes"
+version = "0.0.1"
