+++ components/db_schemes/site01/pyproject.toml
@@ -1,0 +1,7 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "db.site01"
+version = "0.0.1"
